// ClientApp/src/app/models/eo-overdue-kit-work-order-detail.ts
export class EoOverdueKitWorkOrderDetail {
    KEYWORD_DESCRIPTION: string;
    STATION: string;
    VENDOR_NAME: string;
    NCA_VENDOR_NUMBER: number;
    DAYS_LATE: number;
    EO_NUMBER: string;
    DASH_8: string;
    EO_DESC: string;
    QTY_REQ: number;
    REQD_IND: string;
    TTL_OH: number;
    TTL_REQ_EO: number;
    TTL_REQ_30_DAY: number;
    TTL_REQ_60_DAY: number;
    TTL_REQ_90_DAYS: number;
    TTL_REQ_120_DAYS: number;
    TTL_REQ_180_DAYS: number;
    KEY: string;
    COMM: string;
    
}