// ClientApp/src/app/models/eo-overdue-pos-ros-detail.ts
export class EoOverduePosRosDetail {
    KEYWORD: string;
    SHIP_STATION: string;
    VENDOR_NAME: string;
    VENDOR_NBR: number;
    DAYS_LATE: number;
    EO_NUMBER: string;
    DASH_8: string;
    EO_DESC: string;
    QTY_REQ: number;
    REQD_IND: number;
    TTL_OH: number;
    TTL_REQ_EO: number;
    TTL_REQ_30_DAY: number;
    TTL_REQ_60_DAY: number;
    TTL_REQ_90_DAYS: number;
    TTL_REQ_120_DAYS: number;
    TTL_REQ_180_DAYS: number;
    KEY: string;
    COMM: string;
}