// ClientApp/src/app/models/eo-part-in-discrepancy-detail.ts
export class EoPartInDiscrepancyDetail {
    MENBR: string;
    DESCRIPTION: string;
    SCHEDSTA: string;
    FILLSTA: string;
    AR: string;
    QTYREQD: number;
    QTYFILLED: number;
    QTYOPEN: number;
    QTYREQDTODAY: number;
    TOTALREQUIREMENTS: number;
    ALLBOH: number;
    CLTBOH: number;
    PHLBOH: number;
    PITBOH: number;
    PHXBOH: number;
    FILLSTABOH: number;
    POOPENQTY: number;
    ROOPENQTY: number;
    KEY: string;
    COMM: string;
}